[Add: 카페 기능]

+++ src/main/java/cafeSubscription/coffee/domain/cafe/Cafe.java
@@ -14,7 +14,7 @@ public class Cafe {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cafeId;
+    private long cafeId;
 
     @Column(nullable = false)
     private String cafeName;
